expense_tracker/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
crispy-bootstrap5==2024.10
Django==5.1.7
django-crispy-forms==2.3
django-filter==25.1
django-taggit==6.1.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
